chore: rename webcam mod to v4l
